+++ qml/js/bookmarks.js
@@ -22,6 +22,7 @@
 // (no library because variables from the environment are needed)
 
 .import "paths.js" as Paths
+.import harbour.file.browser.Settings 1.0 as Settings
 
 function _defaultFor(arg, val) {
     return typeof arg !== 'undefined' ? arg : val;
@@ -32,11 +33,11 @@ function addBookmark(path, name) {
     path = path.replace(/\/+/g, '/');
     path = path.replace(/\/$/, '');
     name = _defaultFor(name, Paths.lastPartOfPath(path))
-    rawSettings.write("Bookmarks/"+path, name);
+    Settings.RawSettings.write("Bookmarks/"+path, name);
 
     var bookmarks = getBookmarks();
     bookmarks.push(path);
-    rawSettings.write("Bookmarks/Entries", JSON.stringify(bookmarks));
+    Settings.RawSettings.write("Bookmarks/Entries", JSON.stringify(bookmarks));
     main.bookmarkAdded(path);
 }
 
@@ -44,8 +45,8 @@ function removeBookmark(path) {
     if (!path) return;
     var bookmarks = getBookmarks();
     var filteredBookmarks = bookmarks.filter(function(e) { return e !== path; });
-    rawSettings.write("Bookmarks/Entries", JSON.stringify(filteredBookmarks));
-    rawSettings.remove("Bookmarks/"+path);
+    Settings.RawSettings.write("Bookmarks/Entries", JSON.stringify(filteredBookmarks));
+    Settings.RawSettings.remove("Bookmarks/"+path);
     main.bookmarkRemoved(path);
 }
 
@@ -63,19 +64,19 @@ function moveBookmark(path) {
 
     var newIndex = oldIndex === 0 ? bookmarks.length-1 : oldIndex-1;
     bookmarks.splice(newIndex, 0, bookmarks.splice(oldIndex, 1)[0]);
-    rawSettings.write("Bookmarks/Entries", JSON.stringify(bookmarks));
+    Settings.RawSettings.write("Bookmarks/Entries", JSON.stringify(bookmarks));
     main.bookmarkMoved(path);
 }
 
 function hasBookmark(path) {
     if (!path) return false;
-    if (rawSettings.read("Bookmarks/"+path) !== "") return true;
+    if (Settings.RawSettings.read("Bookmarks/"+path) !== "") return true;
     return false;
 }
 
 function getBookmarks() {
     try {
-        var entries = JSON.parse(rawSettings.read("Bookmarks/Entries"));
+        var entries = JSON.parse(Settings.RawSettings.read("Bookmarks/Entries"));
         // remove duplicates
         entries = entries.filter(function(value, index, self){ return self.indexOf(value) === index; });
         return entries;
@@ -83,15 +84,15 @@ function getBookmarks() {
         // The ordering field seems to be empty. It is possible that it was lost because
         // the user moved the settings folder while the app was running. There is no way
         // to restore it, but at least we can make sure all entries are still shown.
-        var keys = rawSettings.keys("Bookmarks").filter(function(e) { return e !== 'Entries'; });
-        rawSettings.write("Bookmarks/Entries", JSON.stringify(keys));
+        var keys = Settings.RawSettings.keys("Bookmarks").filter(function(e) { return e !== 'Entries'; });
+        Settings.RawSettings.write("Bookmarks/Entries", JSON.stringify(keys));
         return keys;
     }
 }
 
 function getBookmarkName(path) {
     if (path === "") return "";
-    var name = rawSettings.read("Bookmarks/"+path);
+    var name = Settings.RawSettings.read("Bookmarks/"+path);
 
     if (name === "") {
         console.warn("empty bookmark name for", path, "- reset to default value");